stagefright: handleSetSurface: Don't dereference null graphic buffers
In metadata mode and when buffers are owned by the native window
, it is possible that the graphics buffer is null.

This is fine since we don't have to attach it to a surface. However,
the logging was trying to dereference this graphic buffer which caused
media to crash sporadically.

This fixes the flakiness in
android.media.cts.MediaPlayerTest#testVideoSurfaceResetting

Change-Id: I89f30ca1177d7118da5a7b0ffbbb91370f9b54ab